Moved the sources to src and lib
