+++ lib/healthcheck/HealthChecker.js
@@ -6,6 +6,7 @@ var HealthChecker = (function () {
         this.router = express.Router();
         this.startTime = Date.now();
         this.services = {};
+        this.isUp = true;
         this.setupRoutes();
     }
     HealthChecker.prototype.register = function (name, checker) {
@@ -21,7 +22,7 @@ var HealthChecker = (function () {
             });
             var status = {
                 services: x,
-                status: "up",
+                status: _this.isUp ? "up" : "down",
                 uptime: uptime
             };
             res.status(200);
