+++ lib/healthcheck/HealthStatus.js
