+++ lib/healthcheck/MongooseServiceChecker.js
