+++ tsconfig.json
@@ -1,12 +1,12 @@
 {
   "compilerOptions": {
-    "outDir": "./js",
+    "outDir": "./lib",
     "allowJs": true,
     // "target": "es5",
     "removeComments": true,
     "sourceMap": true
   },
   "include": [
-    "./ts/**/*"
+    "./src/**/*"
   ]
 }
